network_security_end_to_end/requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.12
pymongo

# -e .